cgl: Fix warnings -Wtautological-constant-out-of-range-compare
In macro cgl_config.m:ADD_ATTR(), cast the compared value to an int
before comparing it to WAFFLE_DONT_CARE (which is itself an int). This
fixes tautological comparisons when the compared value is a bool.

Fixes 2 warnings like the following:

  src/waffle/cgl/cgl_config.m:207:5: warning: comparison of constant
          'WAFFLE_DONT_CARE' (-1) with expression of type 'const bool'
          is always true [-Wtautological-constant-out-of-range-compare]
      ADD_ATTR(kCGLPFASampleBuffers,      attrs->sample_buffers);
      ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  src/waffle/cgl/cgl_config.m:173:21: note: expanded from macro 'ADD_ATTR'
        if ((value) != WAFFLE_DONT_CARE) { \
            ~~~~~~~ ^  ~~~~~~~~~~~~~~~~

+++ src/waffle/cgl/cgl_config.m
@@ -170,7 +170,7 @@ cgl_config_fill_pixel_format_attrs(
     // CGL does not have an analogue for EGL_DONT_CARE. Instead, one indicates
     // "don't care" by omitting the attribute.
     #define ADD_ATTR(name, value) \
-        if ((value) != WAFFLE_DONT_CARE) { \
+        if ((int)(value) != WAFFLE_DONT_CARE) { \
             pixel_attrs[i++] = (name); \
             pixel_attrs[i++] = (value); \
         }
